+++ src/main/java/com/example/swimServer/infrastructure/persistance/maria/swimmer/SwimmerRepository.java
@@ -3,5 +3,5 @@ package com.example.swimServer.infrastructure.persistance.maria.swimmer;
 import com.example.swimServer.domain.model.entity.swimmer.Swimmer;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SwimmerRepository extends CrudRepository<Swimmer, Integer> {
+public interface SwimmerRepository extends CrudRepository<Swimmer, Long> {
 }
